Implement function to get possible next locations
This commit includes the implementation of a function to get a list of
the possible next locations, given a driver's current location.  This uses
hardcoded values, which is not the most scalable technique, but will work'
for the purposes of this assignment.

This commit also includes the tests written for this function in City.java.